feat(github-robot): change code structure
