+++ src/github/github/github.controller.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { GithubController } from './github.controller';
+
+describe('Github Controller', () => {
+  let controller: GithubController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [GithubController],
+    }).compile();
+
+    controller = module.get<GithubController>(GithubController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});
