+++ src/github/github/github.controller.ts
@@ -1,0 +1,17 @@
+import { Controller, Post, Body, Headers } from '@nestjs/common';
+import { IParam, IHeader, Type, IRes } from './github.interfaces';
+import { GithubService } from './github.service';
+
+@Controller('github')
+export class GithubController {
+
+  constructor(private readonly githubService: GithubService) {}
+
+  @Post()
+  async getWebhook(@Body() params: IParam, @Headers() headers: IHeader): Promise<IRes> {
+    console.log('headers', headers)
+    const type: Type = headers['x-github-event'];
+
+    return this.githubService.sendMsg(params, type)
+  }
+}
