+++ src/github/github/github.service.spec.ts
@@ -1,0 +1,18 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { GithubService } from './github.service';
+
+describe('GithubService', () => {
+  let service: GithubService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [GithubService],
+    }).compile();
+
+    service = module.get<GithubService>(GithubService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
